+++ physicsnemo/utils/diffusion/stochastic_sampler.py
@@ -33,9 +33,19 @@ def _apply_wrapper_Cin_channels(patching, input, additional_input=None):
 
 
 @torch.compile()
-def _apply_wrapper_Cout_channels(patching, input, additional_input=None):
+def _apply_wrapper_Cout_channels_no_grad(patching, input, additional_input=None):
     """
-    Apply the patching operation to the input tensor with :math:`C_{out}` channels.
+    Apply the patching operation to an input tensor with :math:`C_{out}`
+    channels that does not require gradients.
+    """
+    return patching.apply(input=input, additional_input=additional_input)
+
+
+@torch.compile()
+def _apply_wrapper_Cout_channels_grad(patching, input, additional_input=None):
+    """
+    Apply the patching operation to an input tensor with :math:`C_{out}`
+    channels that requires gradients.
     """
     return patching.apply(input=input, additional_input=additional_input)
 
@@ -43,6 +53,25 @@ def _apply_wrapper_Cout_channels(patching, input, additional_input=None):
 @torch.compile()
 def _fuse_wrapper(patching, input, batch_size):
     return patching.fuse(input=input, batch_size=batch_size)
+
+
+def _apply_wrapper_select(
+    input: torch.Tensor, patching: GridPatching2D | None
+) -> Callable:
+    """
+    Select the correct patching wrapper based on the input tensor's requires_grad attribute.
+    If patching is None, return the identity function.
+    If patching is not None, return the appropriate patching wrapper.
+    If input.requires_grad is True, return _apply_wrapper_Cout_channels_grad.
+    If input.requires_grad is False, return _apply_wrapper_Cout_channels_no_grad.
+    """
+    if patching:
+        if input.requires_grad:
+            return _apply_wrapper_Cout_channels_grad
+        else:
+            return _apply_wrapper_Cout_channels_no_grad
+    else:
+        return lambda patching, input, additional_input=None: input
 
 
 def stochastic_sampler(
@@ -222,6 +251,10 @@ def stochastic_sampler(
     else:
         patch_embedding_selector = None
 
+    optional_args = {}
+    if lead_time_label is not None:
+        optional_args["lead_time_label"] = lead_time_label
+
     # Main sampling loop.
     x_next = latents * t_steps[0]
     for i, (t_cur, t_next) in enumerate(zip(t_steps[:-1], t_steps[1:])):  # 0, ..., N-1
@@ -235,31 +268,21 @@ def stochastic_sampler(
         # Euler step. Perform patching operation on score tensor if patch-based
         # generation is used denoised = net(x_hat, t_hat,
         # class_labels,lead_time_label=lead_time_label).to(torch.float64)
-        x_hat_batch = (
-            _apply_wrapper_Cout_channels(patching=patching, input=x_hat)
-            if patching
-            else x_hat
+        x_hat_batch = _apply_wrapper_select(input=x_hat, patching=patching)(
+            patching=patching, input=x_hat
         ).to(latents.device)
 
         x_lr = x_lr.to(latents.device)
 
-        if lead_time_label is not None:
-            denoised = net(
-                x_hat_batch,
-                x_lr,
-                t_hat,
-                class_labels,
-                lead_time_label=lead_time_label,
-                embedding_selector=patch_embedding_selector,
-            )
-        else:
-            denoised = net(
-                x_hat_batch,
-                x_lr,
-                t_hat,
-                class_labels,
-                embedding_selector=patch_embedding_selector,
-            )
+        denoised = net(
+            x_hat_batch,
+            x_lr,
+            t_hat,
+            class_labels,
+            embedding_selector=patch_embedding_selector,
+            **optional_args,
+        )
+
         if patching:
             # Un-patch the denoised image
             # (batch_size, C_out, img_shape_y, img_shape_x)
@@ -274,29 +297,19 @@ def stochastic_sampler(
         if i < num_steps - 1:
             # Patched input
             # (batch_size * patch_num, C_out, patch_shape_y, patch_shape_x)
-            x_next_batch = (
-                _apply_wrapper_Cout_channels(patching=patching, input=x_next)
-                if patching
-                else x_next
+            x_next_batch = _apply_wrapper_select(input=x_next, patching=patching)(
+                patching=patching, input=x_next
             ).to(latents.device)
 
-            if lead_time_label is not None:
-                denoised = net(
-                    x_next_batch,
-                    x_lr,
-                    t_next,
-                    class_labels,
-                    lead_time_label=lead_time_label,
-                    embedding_selector=patch_embedding_selector,
-                )
-            else:
-                denoised = net(
-                    x_next_batch,
-                    x_lr,
-                    t_next,
-                    class_labels,
-                    embedding_selector=patch_embedding_selector,
-                )
+            denoised = net(
+                x_next_batch,
+                x_lr,
+                t_next,
+                class_labels,
+                embedding_selector=patch_embedding_selector,
+                **optional_args,
+            )
+
             if patching:
                 # Un-patch the denoised image
                 # (batch_size, C_out, img_shape_y, img_shape_x)